+++ alive_progress/core/configuration.py
@@ -105,7 +105,7 @@ def _format_input_factory(allowed):
 Config = namedtuple('Config', 'title length spinner bar unknown force_tty disable manual '
                               'enrich_print receipt receipt_text monitor elapsed stats '
                               'title_length spinner_length refresh_secs monitor_end '
-                              'elapsed_end stats_end ')
+                              'elapsed_end stats_end ctrl_c')
 
 
 def create_config():
@@ -130,6 +130,7 @@ def create_config():
             title_length=0,
             spinner_length=0,
             refresh_secs=0,
+            ctrl_c=True,
         )
 
     def set_global(theme=None, **options):
@@ -196,6 +197,7 @@ def create_config():
             title_length=_int_input_factory(0, 100),
             spinner_length=_int_input_factory(0, 100),
             refresh_secs=_int_input_factory(0, 60 * 60 * 24),  # maximum 24 hours.
+            ctrl_c=_bool_input_factory(),
             # title_effect=_enum_input_factory(),  # TODO someday.
         )
         assert all(k in validations for k in Config._fields)  # ensures all fields have validations.