+++ alive_progress/utils/terminal/tty.py
@@ -11,18 +11,20 @@ def get(original_stdout):
         # more resilient one, although 7x slower than os' one.
         return shutil.get_terminal_size()[0]
 
-    def _ansi_escape_code(sequence, param=''):
+    def _ansi_escape_sequence(code, param=''):
         def inner(_available=None):  # because of jupyter.
-            write(text)
+            write(inner.sequence)
 
-        text = f'\x1b[{param}{sequence}'
+        inner.sequence = f'\x1b[{param}{code}'
         return inner
 
-    clear_line = _ansi_escape_code('2K\r')  # clears the entire line: CSI n K -> with n=2.
-    clear_end = _ansi_escape_code('K')  # clears line from cursor: CSI K.
-    hide_cursor = _ansi_escape_code('?25l')  # hides the cursor: CSI ? 25 l.
-    show_cursor = _ansi_escape_code('?25h')  # shows the cursor: CSI ? 25 h.
-    factory_cursor_up = lambda num: _ansi_escape_code('A', num)  # sends cursor up: CSI {x}A.
+    def factory_cursor_up(num):
+        return _ansi_escape_sequence('A', num)  # sends cursor up: CSI {x}A.
+
+    clear_line = _ansi_escape_sequence('2K\r')  # clears the entire line: CSI n K -> with n=2.
+    clear_end = _ansi_escape_sequence('K')  # clears line from cursor: CSI K.
+    hide_cursor = _ansi_escape_sequence('?25l')  # hides the cursor: CSI ? 25 l.
+    show_cursor = _ansi_escape_sequence('?25h')  # shows the cursor: CSI ? 25 h.
     carriage_return = '\r'
 
     return SimpleNamespace(**locals())