+++ alive_progress/utils/terminal/void.py
@@ -6,18 +6,18 @@ def flush():
     pass
 
 
-def _ansi_escape_code(_=''):
+def _ansi_escape_sequence(_=''):
     def inner(_=None):
         pass
 
     return inner
 
 
-clear_line = _ansi_escape_code()
-clear_end = _ansi_escape_code()
-hide_cursor = _ansi_escape_code()
-show_cursor = _ansi_escape_code()
-factory_cursor_up = lambda _: _ansi_escape_code()
+clear_line = _ansi_escape_sequence()
+clear_end = _ansi_escape_sequence()
+hide_cursor = _ansi_escape_sequence()
+show_cursor = _ansi_escape_sequence()
+factory_cursor_up = lambda _: _ansi_escape_sequence()
 
 
 def cols():